♻️refactor: solicitations accordions
